Use PF toggle instead of dropdown for all services dropdown.
