+++ src/components/AllServicesDropdown/AllServicesDropdown.scss
@@ -1,5 +1,12 @@
 @import '~@patternfly/patternfly/sass-utilities/_all';
 
+.chr-c-link-service-toggle {
+  // override menu toggle extra bottom border
+  &.pf-c-menu-toggle::after {
+    border-bottom: none;
+  }
+}
+
 .chr-c-page__services-nav-dropdown-menu {
   position: absolute;
   top: 70px;
@@ -7,6 +14,7 @@
   left: 0;
   right: 0;
   height: calc(100vh - 70px);
+  width: 100% !important; // override popper calculated width
   overflow-y: auto;
   overflow-x: none;
   .pf-c-dropdown__menu::before {
