+++ src/components/Header/Header.tsx
@@ -21,7 +21,7 @@ import { ReduxState } from '../../redux/store';
 import { activationRequestURLs } from '../../utils/consts';
 import { isBeta, isFedRamp, isProd } from '../../utils/common';
 import SearchInput from '../Search/SearchInput';
-import ServicesNewNav from '../../layouts/ServicesNewNav';
+import AllServicesDropdown from '../AllServicesDropdown/AllServicesDropdown';
 // import ServicesNewNav from '../../layouts/ServicesNewNav';
 
 const FeedbackRoute = ({ user }: { user: DeepRequired<ChromeUser> }) => {
@@ -62,7 +62,7 @@ export const Header = () => {
                   ) : (
                     <>
                       {/* <ServicesLink /> */}
-                      <ServicesNewNav />
+                      <AllServicesDropdown />
                       <FavoritesLink />
                     </>
                   )}
